Create NegativeOnOneSide.cpp
This Program takes input as an array of size n. The value of n is to be input of the user. The program then brings all the negative numbers of the array to one side.
It uses vectors which are present in the C++ standard template library.  It also uses a array data structure., For loops.